Rename static-pages -> legal-pages
