+++ src/legal-pages/config.ts
