+++ src/legal-pages/index.tsx
@@ -48,7 +48,7 @@ import {
 
 const defaultLanguage = Instance.En
 
-export async function staticPages(
+export async function legalPages(
   request: Request,
   unrevisedConfig = defaultUnrevisedConfig,
   revisedConfig = defaultRevisedConfig
@@ -324,8 +324,10 @@ async function fetchContent<A extends Page>(
     const text = await response.text()
     const rawContent = page.url.endsWith('.md') ? markdownToHtml(text) : text
     const sanitizedContent = sanitizeHtml(rawContent)
-    const content = sanitizedContent
-      .replace('JS-GOOGLE-ANALYTICS-DEACTIVATE', 'javascript:gaOptout();')
+    const content = sanitizedContent.replace(
+      'JS-GOOGLE-ANALYTICS-DEACTIVATE',
+      'javascript:gaOptout();'
+    )
 
     return { ...page, content }
   } else {
